Changes && Correction - 24-07-2026

+++ verificationFile/documentation/sign_info_doc_upload.php
@@ -8,7 +8,7 @@ $signType_relationship = $_POST['signType_relationship'];
 $doc_Count             = $_POST['doc_Count'];
 $cus_profile_id        = $_POST['doc_cus_profile_id'];
 $signedID              = $_POST['signedID'];
-$filesArr3             = $_FILES['signdoc_upd'];
+$filesArr3             = $_FILES['signdoc_upd'] ?? '';
 
 if ($sign_type == '1') {
     $qry = $connect->query("SELECT fam.id from verification_family_info fam JOIN customer_profile cp on cp.guarentor_name = fam.id where cp.req_id = $req_id");
@@ -24,25 +24,40 @@ if ($signedID == '') {
     $update = $connect->query("UPDATE `signed_doc_info` SET `cus_id`='$cus_id',`doc_name`='0',`sign_type`='$sign_type',`signType_relationship`='$signType_relationship',`doc_Count`='$doc_Count' WHERE `id`='$signedID' ");
 }
 
+if(!empty($filesArr3['name'])){
+    $filePath = "../../uploads/verification/signed_doc/";
 
-$connect->query("DELETE FROM `signed_doc` WHERE `signed_doc_id` ='$signedID'");
+    // Get all files for this signed_doc_id
+    $stmt = $connect->query("SELECT upload_doc_name FROM signed_doc WHERE signed_doc_id = '$signedID'");
+    $files = $stmt->fetchAll(PDO::FETCH_ASSOC);
 
-foreach ($filesArr3['name'] as $key => $val) {
-    $fileName = basename($filesArr3['name'][$key]);
-    $targetFilePath = "../../uploads/verification/signed_doc/" . $fileName;
+    // Delete files from folder
+    foreach ($files as $fileinfo) {
+        $file = $filePath . $fileinfo['upload_doc_name'];
 
-    $fileExtension = pathinfo($targetFilePath, PATHINFO_EXTENSION);
-
-    $uniqueFileName = uniqid() . '.' . $fileExtension;
-    while (file_exists("../../uploads/verification/signed_doc/" . $uniqueFileName)) {
-        $uniqueFileName = uniqid() . '.' . $fileExtension;
+        if (!empty($fileinfo['upload_doc_name']) && file_exists($file)) {
+            unlink($file);
+        }
     }
 
-    if (move_uploaded_file($filesArr3["tmp_name"][$key], "../../uploads/verification/signed_doc/" . $uniqueFileName)) {
-        $update =  $connect->query("INSERT INTO `signed_doc`(`cus_id`,`req_id`,`signed_doc_id`, `upload_doc_name`) VALUES ('$cus_id','$req_id','$signedID','$uniqueFileName')");
+    $connect->query("DELETE FROM `signed_doc` WHERE `signed_doc_id` ='$signedID'");
+
+    foreach ($filesArr3['name'] as $key => $val) {
+        $fileName = basename($filesArr3['name'][$key]);
+        $targetFilePath = $filePath . $fileName;
+
+        $fileExtension = pathinfo($targetFilePath, PATHINFO_EXTENSION);
+
+        $uniqueFileName = uniqid() . '.' . $fileExtension;
+        while (file_exists($filePath . $uniqueFileName)) {
+            $uniqueFileName = uniqid() . '.' . $fileExtension;
+        }
+
+        if (move_uploaded_file($filesArr3["tmp_name"][$key], $filePath . $uniqueFileName)) {
+            $update =  $connect->query("INSERT INTO `signed_doc`(`cus_id`,`req_id`,`signed_doc_id`, `upload_doc_name`) VALUES ('$cus_id','$req_id','$signedID','$uniqueFileName')");
+        }
     }
 }
-
 
 if ($update) {
     $result = "Signed Doc Info Uploaded Successfully.";
